Restructure dictionary modules
