+++ vint/ast/dictionary/abbreviations.py
