+++ vint/linting/policy/prohibit_abbreviation_option.py
@@ -3,7 +3,7 @@ from vint.ast.node_type import NodeType
 from vint.linting.level import Level
 from vint.linting.policy.abstract_policy import AbstractPolicy
 from vint.linting.policy_registry import register_policy
-from vint.linting.policy.abbreviation import (
+from vint.ast.dictionary.abbreviations import (
     Abbreviations,
     AbbreviationsIncludingInvertPrefix,
 )
